refactor : 공통 컴포넌트 Thumbnail 추가 및 기존에 중복되었던 썸네일 컴포넌트 대신 공통 컴포넌트 배치
